# openapi.json
{
  "components": {
    "schemas": {
...
      "FieldProblem": {
        "description": "One field-level validation message, as the 422 handler emits them.",
        "properties": {
          "field": {
            "description": "Dotted path including the source, e.g. body.title",
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "FieldProblem",
        "type": "object"
      },
...
      "IterationResponse": {
        "properties": {
          "closed": {
            "title": "Closed",
            "type": "boolean"
          },
          "ends_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends On"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "starts_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          }
        },
        "required": [
          "id",
          "name",
          "starts_on",
          "ends_on",
          "closed"
        ],
        "title": "IterationResponse",
        "type": "object"
      },
      "LabelResponse": {
        "properties": {
          "color": {
            "title": "Color",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "color"
        ],
        "title": "LabelResponse",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 ``application/problem+json`` — **every** error response.\n\nDeclared as a model, and attached to the routers, for a reason that is easy to\nmiss: the exception handlers are invisible to FastAPI's schema generation, so\nwithout this the OpenAPI document would describe only the happy paths. An\nintegrator reading ``/docs`` would then have to *discover* the error shape by\ncausing errors, which is precisely the \"first thing an integrator hits\" §8.6\nis about. It also means the shape lands in API-5's snapshot, so changing it\nbecomes a visible contract change.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldProblem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present on a 422; one entry per rejected field.",
            "title": "Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Human-readable, and names the next step.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable URI identifying the error class. **Match on this**, never on ``title`` — the title is human-readable prose and may be reworded.",
            "examples": [
              "https://relay.internal/problems/not_found"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "UserResponse": {
        "description": "Id and display name. See the module note on why there is no email here.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name"
        ],
        "title": "UserResponse",
        "type": "object"
      },
...
      "relay__api__v1__meta__TicketFieldResponse": {
        "properties": {
          "domain_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Scope"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "visible": {
            "title": "Visible",
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "label",
          "type",
          "domain_scope",
          "visible"
        ],
        "title": "TicketFieldResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Relay",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/meta/iterations": {
      "get": {
        "operationId": "iterations_api_v1_meta_iterations_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_closed",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Closed",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IterationResponse"
                  },
                  "title": "Response Iterations Api V1 Meta Iterations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Iterations",
        "tags": [
          "meta (v1)"
        ]
      }
    },
    "/api/v1/meta/labels": {
      "get": {
        "operationId": "labels_api_v1_meta_labels_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LabelResponse"
                  },
                  "title": "Response Labels Api V1 Meta Labels Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Labels",
        "tags": [
          "meta (v1)"
        ]
      }
    },
    "/api/v1/meta/ticket-fields": {
      "get": {
        "description": "The tenant's ``ai_context`` schema (TKT-2 · §7.3).\n\nThis is what makes ``ai_context`` writable by an external system at all: the\nwrite path validates against exactly this config, so a consumer that reads it\nfirst knows which keys will be accepted instead of discovering it from a 422.",
        "operationId": "ticket_fields_api_v1_meta_ticket_fields_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/relay__api__v1__meta__TicketFieldResponse"
                  },
                  "title": "Response Ticket Fields Api V1 Meta Ticket Fields Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Ticket Fields",
        "tags": [
          "meta (v1)"
        ]
      }
    },
    "/api/v1/meta/users": {
      "get": {
        "description": "Members, for resolving an assignee. **No emails, and no handles either.**\n\nThe web directory returns the mention handle because people type it; this one\ndoes not, because the handle is the local part of a work address and handing a\nmachine principal the shape of everyone's email is the leak §8.3 is guarding\nagainst.",
        "operationId": "users_api_v1_meta_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "title": "Response Users Api V1 Meta Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Users",
        "tags": [
          "meta (v1)"
        ]
      }
    },
...
  }
}